fix(virtual-screens): address PR #247 review findings — crashes, leaks, correctness
Critical: fix window leak on physical screen removal (overlayservice),
add null checks for m_windowTrackingAdaptor in handleRestore and snap
assist handler to prevent crashes during startup/teardown races.

Major: extract DRY helpers for prefix-erase cleanup (10x duplication),
fix useShaderForScreen resolving physical instead of virtual screen ID,
update screenadaptor cached effective IDs on VS config change, fix
VirtualScreenDef operator== to use exact comparison (was fuzzy/intransitive).

Minor: rename screenName→screenId in D-Bus JSON for consistency, use
RestoreSentinel constant, consolidate FallbackScreen dimensions into
constants.h, narrow drag guard to only skip dragged window, use
constexpr QLatin1StringView for effect D-Bus constants, replace magic
numbers with named constants, add missing isSticky to buildStateObject,
scale divider width with Kirigami.Units.

+++ src/daemon/overlayservice.h
@@ -401,7 +401,7 @@ private:
     // Appended to each configureLayerSurface() scope so KWin sees every
     // new surface as unique, avoiding configure rate-limiting after rapid
     // destroy/recreate cycles on Vulkan.
-    int m_scopeGeneration = 0;
+    uint64_t m_scopeGeneration = 0;
 
     // CAVA audio visualization
     std::unique_ptr<CavaService> m_cavaService;
